PGNTCORE-24 correzione gestione bigint

+++ src/com/seda/payer/core/bean/Gruppo.java
@@ -121,7 +121,7 @@ public class Gruppo extends Lifecycle implements Serializable{
 	 * @see com.seda.payer.commons.bean.Lifecycle#onUpdate(java.sql.CallableStatement)
 	 */
 	public void onUpdate(CallableStatement arg) throws SQLException {
-		arg.setInt(1, this.chiaveGruppo.intValue());	
+		arg.setLong(1, this.chiaveGruppo.longValue());
 		arg.setString(2, this.codiceGruppo);	
 		arg.setString(3, this.descrizioneLinguaItaliana);	
 		arg.setString(4, this.descrizioneAltraLingua);
